cleanup hashFiles step
